added indentation levels and prompts for brace blocks
