+++ src/globals.hpp
@@ -3,11 +3,10 @@
 #include <curses.h>
 #include <stack>
 
-extern stack<pair<string, int>> HISTORY;
-extern stack<pair<string, int>> MEMORY;
+extern stack<pair<string, int>> HISTORY, MEMORY;
 extern WINDOW* PAD;
-extern unsigned int CURSOR_X;
-extern unsigned int CURSOR_Y;
-extern unsigned int MIN_X;
-extern unsigned int MAXLINE_X;
+extern unsigned int CURSOR_X, CURSOR_Y;
+extern unsigned int MIN_X, MAXLINE_X;
 extern int PADPOS;
+extern int WIN_HEIGHT, WIN_WIDTH;
+extern int INDENT_LEVEL, INDENT_SPACES;
